DoodleViewContorller의 내용을 파일단위 분리
UICollectionViewDelegateFlowLayout을 채택한 class 생성하여, delegate 연결.

DataSource내용은 extension 유지 but, Pseudo 접두어 파일로 분리하여 관리의 용이를 꾀함
