+++ PhotosApp/PhotosApp/DoodleViewComponents/DoodleViewController.swift
@@ -10,6 +10,7 @@ import Photos
 
 class DoodleViewController: UICollectionViewController, UIGestureRecognizerDelegate {
     var doodles = [Doodle]()
+    let delegateFlowLayout = DoodleViewDelegateFlowLayout()
     
     override var canBecomeFirstResponder: Bool {
         return true
@@ -18,13 +19,12 @@ class DoodleViewController: UICollectionViewController, UIGestureRecognizerDeleg
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        self.collectionView.delegate = delegateFlowLayout
+        
         self.title = "Doodle"
-
-        self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "close", style: .plain, target: self, action: #selector(closeButtonPushed))
         self.collectionView.backgroundColor = .darkGray
-        
+        self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "close", style: .plain, target: self, action: #selector(closeButtonPushed))
         loadDoodles()
-        
     }
     
     func loadDoodles() {
@@ -39,76 +39,5 @@ class DoodleViewController: UICollectionViewController, UIGestureRecognizerDeleg
     
     @objc func closeButtonPushed() {
         self.dismiss(animated: true, completion: nil)
-    }
-}
-
-extension DoodleViewController {
-    override func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return doodles.count
-    }
-    
-    override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "MyCell", for: indexPath) as? CustomCell
-        else {
-            return UICollectionViewCell()
-        }
-        
-        DispatchQueue.global().async {
-            let doodle = self.doodles[indexPath.row]
-            let imageURL = doodle.image
-            guard let url = URL(string: imageURL) else { return }
-            var data = Data()
-            do {
-                data = try Data(contentsOf: url)
-            } catch {
-                return
-            }
-            DispatchQueue.main.async {
-                cell.imageView.image = UIImage(data: data)
-            }
-        }
-        
-        setTouchRecognizer(To: cell, indexPathRow: indexPath.row)
-        
-        return cell
-    }
-    
-    func setTouchRecognizer(To cell : CustomCell, indexPathRow : Int) {
-        let longPressRecognizer = UITapGestureRecognizer(target: self, action: #selector(longPressedCell(gesture:)))
-        longPressRecognizer.cancelsTouchesInView = false
-        cell.imageView.tag = indexPathRow // add this
-        longPressRecognizer.numberOfTapsRequired = 1 // add this
-        longPressRecognizer.delegate = self
-        cell.imageView.isUserInteractionEnabled = true  // add this
-        cell.imageView.addGestureRecognizer(longPressRecognizer)
-    }
-    
-    @objc func longPressedCell(gesture: UITapGestureRecognizer) {
-        guard let gestureView = gesture.view else { return }
-        let indexPath = NSIndexPath(row: gestureView.tag, section: 0)
-        
-        guard let cell = collectionView.cellForItem(at: indexPath as IndexPath) as? CustomCell else { return }
-        cell.becomeFirstResponder()
-        
-        guard let cellImage = cell.imageView.image else { return }
-        PhotoManager.shared.doodleClipBoardImage = cellImage
-        
-        if let rectView = gesture.view {
-            let saveMenuItem = UIMenuItem(title: "Save💾", action: #selector(save))
-            UIMenuController.shared.menuItems = [saveMenuItem]
-            UIMenuController.shared.arrowDirection = .default
-            UIMenuController.shared.showMenu(from: rectView, rect: rectView.frame)
-        }
-    }
-    
-    @objc func save() {
-        let image = PhotoManager.shared.doodleClipBoardImage
-        UIImageWriteToSavedPhotosAlbum(image, nil, nil, nil)
-    }
-}
-
-extension DoodleViewController : UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: 110, height: 50)
     }
 }
